LPD-29276 Rename rolesUtil file
